Split functions that release the runtime lock into separate module
In preparation for introducing a printing function that should
really release the runtime lock.

Allocating/freeing the structure is moved into the module that releases
the runtime lock, whereas the value recording and query functions don't
release it.

Just read in the manual that Dune supports multiple 'function_description'
entries with different concurrency settings.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>